Adding more test cases
